Move witness dispatching to ExtractedWitnessInvariant
